requirements.txt
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.7
websockets==12.0
